Feat: Create stat data class and pass spec and update spec_helper
