+++ lib/stat_data.rb
@@ -1,0 +1,34 @@
+require 'csv'
+require_relative 'game'
+require_relative 'team'
+require_relative 'game_team'
+
+class StatData
+  attr_reader :game_data, :team_data, :game_teams_data
+
+  def initialize(locations)
+    @game_data = CSV.read locations[:games], headers: true, header_converters: :symbol
+    @team_data = CSV.read locations[:teams], headers: true, header_converters: :symbol
+    @game_teams_data = CSV.read locations[:game_teams], headers: true, header_converters: :symbol
+  end
+
+  def all_games
+    @game_data.map do |row|
+      Game.new(row)
+    end
+  end
+
+  def all_teams
+    @team_data.map do |row|
+      team = Team.new(row)
+      team.games = all_games.select { |game| game.home_id == team.team_id || game.away_id == team.team_id }
+      team
+    end
+  end
+
+  def all_game_teams
+    @game_teams_data.map do |row|
+      GameTeam.new(row)
+    end
+  end
+end
