+++ spec/spec_helper.rb
@@ -9,3 +9,4 @@ require './lib/league'
 require './lib/game'
 require './lib/season'
 require './lib/game_team'
+require './lib/stat_data'
